+++ CHANGES.md
@@ -1,5 +1,12 @@
 # Changes and migration requirements
 
+## Version 0.0.5
+
+* Non-breaking space (\u00A0) treated like other whitespace
+* `parse_repeat_phrase()` implemented for some simple phrases.
+* `parse()` can return whitespace tokens, when optional arg `ignore_whitespace`
+  is `False`
+
 ## Version 0.0.4
 
 * Better guessing of the year